+++ RELEASES.md
@@ -7,13 +7,18 @@
 The bottom of "About", and the top of "tests", should say:
 ```
 @ki1r0y/distributed-security 1.2.4
-@kilroy-code/flexstore 0.0.53
-@kilroy-code/ui-components 0.0.39
+@kilroy-code/flexstore 0.0.54
+@kilroy-code/ui-components 0.0.40
 ```
 
-and the "About" page additionally says `Fairshare 0.3x.0`. (See next.)
+and the "About" page additionally says `Fairshare 0.3.2`. (See next.)
 
 If it doesn't, then please reload twice. Why twice? The app is designed to start up quickly from cached files -- even if the page server goes away or you do not have Internet connectivity. But reload also checks for new versions to put in the cache. You won't see those new versions until the _next_ reload.
+
+## 0.4.0
+- Check for synchronization before making new records, giving the user the option to go to Relays screen if necessary.
+- Do not merge bogus versions.
+- Fix "No viable record found" so that it gives option of bug report.
 
 ## 0.3x.1
 - Improve stale source and data handling.